pyproject.toml
[project]
name = "ai-bootcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.20",
    "ipykernel>=6.30.1",
    "numpy>=2.3.2",
    "openai-agents>=0.2.8",
    "openai>=1.100.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
